requirements.txt
atomicwrites==1.1.5
attrs==18.1.0
certifi==2018.4.16
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
coverage==4.5.3
defusedxml==0.5.0
dj-database-url==0.5.0
Django==2.2.3
django-allauth==0.39.1
django-filter==1.1.0
django-rest-auth==0.9.5
djangorestframework==3.8.2
djangorestframework-jwt==1.11.0
factory-boy==2.10.0
Faker==0.8.16
gunicorn==19.8.1
idna==2.7
itypes==1.1.0
Jinja2==2.10
Markdown==2.6.11
MarkupSafe==1.0
more-itertools==4.2.0
oauthlib==2.0.7
pluggy==0.6.0
psycopg2==2.7.5
py==1.5.4
PyJWT==1.6.1
pytest==3.5.0
pytest-cov==2.5.1
pytest-django==3.1.2
python-dateutil==2.7.3
python-decouple==3.1
python3-openid==3.1.0
pytz==2018.5
redis==2.10.6
requests==2.19.1
requests-oauthlib==0.8.0
six==1.11.0
sqlparse==0.3.0
text-unidecode==1.2
uritemplate==3.0.0
urllib3==1.23
urlwait==0.4
whitenoise==3.3.1
